[AIRFLOW-4233] Remove Template Extension from Bq to GCS Operator (#5456)
The issue happens when we have destination url ending with .sql.
It is related to the defined template extension template_ext = ('.sql',)

The operator looks for jinja template, however, it's an output path, so the
file is not found when looking for any jinja template syntax.

The BigQueryToCloudStorageOperator doesn't have sql parameter as it doesn't
work with queries at all. It takes only tables.

(cherry picked from commit 250131d66fd876521a4254d75a1a16954f5488a8)