[AIRFLOW-4871] Allow creating DagRuns via RBAC UI (#5507)
All this needed to enable the form was to add `can_add` to the
permission list

The run_id field is a required form (though the DB doesn't have it as
not nullable) - the scheduler requires it. so I have enabled the
required validation for it.

The `validators_columns` attribute on the view was ignored by FAB
because we set `add_form` and `edit_form` directly, so I have removed
the property

(cherry picked from commit cbb1494397e5e27c672bcbcc5018882253ac273d)