Add custom patches
Configmap mounts aren't needed anymore as the latest airflow version
has handling for the `airflow_local_settings.py` file directly in the
`airflow.cfg` file. I'll verify that works before this goes out.

The mounts for our gcloud secrets objects are still needed, so adding
those blocks back in.

One thing that is notably missing is the manual setting of resources
for the worker pods. We used to hijack this to set the defaults for
CPU and memory, but the module has been rewritten and that "spot"
isn't there anymore. I'm going to see what happens when we deploy
stage without the default values to see how it behaves.